server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
opencv-python
scikit-learn
python-dotenv
svgwrite